requirements.txt
pytorch==0.4.1
torchvision==0.2.1

